* removed warnings

+++ xlators/performance/read-ahead/src/read-ahead.h
@@ -96,13 +96,13 @@ struct ra_file {
   int32_t refcount;
   pthread_mutex_t file_lock;
   struct stat stbuf;
-  int32_t page_size;
-  int32_t page_count;
+  size_t page_size;
+  uint32_t page_count;
 };
 
 struct ra_conf {
   size_t page_size;
-  int32_t page_count;
+  uint32_t page_count;
   void *cache_block;
   struct ra_file files;
   char force_atime_update;
